fix tests, allow imports for parameters_lookup in deprecated module
