+++ beep/features/core.py
@@ -6,7 +6,7 @@ from scipy.interpolate import interp1d
 from beep import PROTOCOL_PARAMETERS_DIR
 from beep.features import featurizer_helpers
 from beep.features.base import BEEPFeaturizer, BEEPFeaturizationError
-from beep.utils.parameters_lookup import get_protocol_parameters
+from beep.deprecated.parameters_lookup import get_protocol_parameters
 from beep.structure.base import get_CV_segment_from_charge
