+++ beep/utils/__init__.py
@@ -2,7 +2,6 @@ import os
 import hashlib
 import json
 from collections import OrderedDict
-from .parameters_lookup import get_protocol_parameters, get_project_sequence, get_diagnostic_parameters
 from .splice import MaccorSplice
 from pydash import get, set_with, unset, merge
